refactor + clenaup
