+++ packages/packages/libs/editor-mcp/src/adapters/angie-adapter.ts
@@ -1,20 +1,27 @@
 import { type AngieMcpSdk } from '@elementor-external/angie-sdk';
 
-import { getRegisteredMcpServers, toMCPTitle } from '../mcp-registry';
+import {
+	type getRegisteredMcpServers as _getRegisteredMcpServers,
+	type toMCPTitle as _toMCPTitle,
+} from '../mcp-registry';
 import { type IMcpRegistrationAdapter } from './types';
 
 const MAX_RETRIES = 3;
 
 export class AngieMcpAdapter implements IMcpRegistrationAdapter {
-	constructor( private readonly sdk: AngieMcpSdk ) {}
+	constructor(
+		private readonly sdk: AngieMcpSdk,
+		private readonly getRegisteredMcpServers: typeof _getRegisteredMcpServers,
+		private readonly toMCPTitle: typeof _toMCPTitle
+	) {}
 
 	async activate(): Promise< void > {
 		await this.sdk.waitForReady();
-		await this.registerEntries( getRegisteredMcpServers(), MAX_RETRIES );
+		await this.registerEntries( this.getRegisteredMcpServers(), MAX_RETRIES );
 	}
 
 	private async registerEntries(
-		entries: ReturnType< typeof getRegisteredMcpServers >,
+		entries: ReturnType< typeof _getRegisteredMcpServers >,
 		retry: number
 	): Promise< void > {
 		if ( retry === 0 ) {
@@ -30,7 +37,7 @@ export class AngieMcpAdapter implements IMcpRegistrationAdapter {
 		for ( const [ key, mcpServer, description ] of entries ) {
 			try {
 				await this.sdk.registerLocalServer( {
-					title: toMCPTitle( key ),
+					title: this.toMCPTitle( key ),
 					name: `editor-${ key }`,
 					server: mcpServer,
 					version: '1.0.0',
