+++ packages/packages/libs/editor-mcp/src/init.ts
@@ -1,21 +1,4 @@
-import { AngieMcpAdapter } from './adapters/angie-adapter';
-import { type ModelContext, WebMCPAdapter } from './adapters/web-mcp-adapter';
-import { activateAdapters, registerMcpAdapter, signalMcpReady } from './mcp-registry';
-import { getSDK } from './utils/get-sdk';
-import { isAngieAvailable } from './utils/is-angie-available';
-
-type ModelContextHost = {
-	modelContext?: ModelContext;
-};
-
-function getModelContext(): ModelContext | undefined {
-	const documentModelContext =
-		typeof document !== 'undefined' ? ( document as unknown as ModelContextHost ).modelContext : undefined;
-	const navigatorModelContext =
-		typeof navigator !== 'undefined' ? ( navigator as unknown as ModelContextHost ).modelContext : undefined;
-
-	return documentModelContext || navigatorModelContext;
-}
+import { createAndRegisterAdapters, signalMcpReady } from './mcp-registry';
 
 let isInitialized = false;
 export function startMCPServer() {
@@ -23,17 +6,8 @@ export function startMCPServer() {
 		return;
 	}
 	isInitialized = true;
-	const modelContext = getModelContext();
 
-	if ( modelContext ) {
-		registerMcpAdapter( new WebMCPAdapter( modelContext ) );
-	}
-
-	if ( isAngieAvailable() ) {
-		registerMcpAdapter( new AngieMcpAdapter( getSDK() ) );
-	}
-
-	activateAdapters();
+	createAndRegisterAdapters();
 	signalMcpReady();
 }
